+++ common/test/src/io/tst_FgdParser.cpp
@@ -17,11 +17,11 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "TestParserStatus.h"
 #include "el/ELTestUtils.h"
 #include "io/DiskIO.h"
 #include "io/FgdParser.h"
 #include "io/Reader.h"
-#include "io/TestParserStatus.h"
 #include "io/TraversalMode.h"
 #include "mdl/EntityDefinition.h"
 #include "mdl/EntityDefinitionTestUtils.h"
