+++ common/test/src/io/tst_ImageFileSystem.cpp
@@ -17,6 +17,7 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "Matchers.h"
 #include "TestUtils.h"
 #include "io/DiskIO.h"
 #include "io/DkPakFileSystem.h"
@@ -29,7 +30,6 @@
 #include <filesystem>
 
 #include "catch/CatchConfig.h"
-#include "catch/Matchers.h"
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators.hpp>
