+++ common/test/src/io/tst_NodeWriter.cpp
@@ -17,6 +17,7 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "Matchers.h"
 #include "TestUtils.h"
 #include "io/NodeWriter.h"
 #include "mdl/BrushBuilder.h"
@@ -41,7 +42,6 @@
 #include <vector>
 
 #include "catch/CatchConfig.h"
-#include "catch/Matchers.h"
 
 #include <catch2/catch_test_macros.hpp>
