+++ common/test/src/io/tst_TestFileSystem.cpp
@@ -17,6 +17,7 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "Matchers.h"
 #include "TestFileSystem.h"
 #include "io/FileSystem.h"
 #include "io/TraversalMode.h"
@@ -29,7 +30,6 @@
 #include <filesystem>
 
 #include "catch/CatchConfig.h"
-#include "catch/Matchers.h"
 
 #include <catch2/catch_test_macros.hpp>
