+++ common/test/src/mdl/tst_BrushNode.cpp
@@ -17,9 +17,9 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "TestParserStatus.h"
 #include "TestUtils.h"
 #include "io/NodeReader.h"
-#include "io/TestParserStatus.h"
 #include "mdl/BezierPatch.h"
 #include "mdl/BrushBuilder.h"
 #include "mdl/BrushFace.h"
@@ -334,7 +334,7 @@ TEST_CASE("BrushNode (Regression)", "[regression]")
 {
   auto taskManager = kdl::task_manager{};
   const auto worldBounds = vm::bbox3d{8192.0};
-  auto status = io::TestParserStatus{};
+  auto status = TestParserStatus{};
 
   SECTION("1186")
   {
