+++ common/test/src/mdl/tst_Map_Geometry.cpp
@@ -19,6 +19,7 @@
 
 
 #include "MapFixture.h"
+#include "Matchers.h"
 #include "PreferenceManager.h"
 #include "Preferences.h"
 #include "TestFactory.h"
