+++ lib/TbBaseLib/CMakeLists.txt
@@ -19,3 +19,4 @@ target_link_libraries(TbBaseLib PRIVATE CompilerConfig PrecompileStdHeaders)
 target_link_libraries(TbBaseLib PUBLIC KdLib VmLib fmt::fmt-header-only)
 
 add_subdirectory(test)
+add_subdirectory(test-utils)
