+++ lib/TbBaseLib/test-utils/README.md
@@ -1,0 +1,3 @@
+# TrenchBroom Base Test Utility Library
+
+Contains a few generic helpers for writing tests such as Catch matchers and string makers.
