+++ lib/TbBaseLib/test-utils/include/TestParserStatus.h
@@ -1,0 +1,50 @@
+/*
+ Copyright (C) 2010 Kristian Duske
+
+ This file is part of TrenchBroom.
+
+ TrenchBroom is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ TrenchBroom is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include "Logger.h"
+#include "ParserStatus.h"
+
+#include <map>
+#include <string>
+#include <vector>
+
+namespace tb
+{
+
+class TestParserStatus : public ParserStatus
+{
+private:
+  static NullLogger _logger;
+  std::map<LogLevel, std::vector<std::string>> m_messages;
+
+public:
+  TestParserStatus();
+
+public:
+  size_t countStatus(LogLevel level) const;
+  const std::vector<std::string>& messages(LogLevel level) const;
+
+private:
+  void doProgress(double progress) override;
+  void doLog(LogLevel level, const std::string& str) override;
+};
+
+} // namespace tb
