+++ lib/TbBaseLib/test-utils/test/.clang-tidy
@@ -1,0 +1,3 @@
+InheritParentConfig: true
+Checks: >
+  -readability-container-size-empty,
